Add PaymentIntentId to Orders table
Introduced a new column `PaymentIntentId` to the `Orders` table in the database schema. Updated the `Order` entity in `StoreDbContextModelSnapshot.cs` to include this new property with a column type of `nvarchar(max)`. Created a migration `20251127211445_AddPaymentIntentIdColumnToBasket` to handle the addition and potential removal of this column. Generated the corresponding designer file to reflect the updated model state.